feat(task): Add core task infrastructure and traits
Implement foundational task management system with key
components:
- Introduce `TaskState` for internal task-local storage
- Define `Runnable` trait for task logic implementation
- Create `Task` trait for task spawning and message handling
- Add module structure for task-related functionality
